requirements.txt
streamlit==1.37.1
streamlit-authenticator==0.3.3
PyYAML==6.0.1
python-docx==1.1.2
pysqlite3-binary 